% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_income}
\alias{clean_income}
\title{Utility functions}
\usage{
clean_income(
  x,
  w = rep(1, length(x)),
  no.negatives = FALSE,
  no.nans = TRUE,
  no.infinites = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{an income vector or matrix}

\item{w}{the vector of weights}

\item{no.negatives}{a logical indicating whether negative values in the data
should be removed. The default value is false}

\item{no.nans}{a logical indicating whether NaN:s in the data should be
removed. The default value is TRUE}

\item{no.infinites}{a logical indicating whether infinite values in the data
should be removedd. The default value is TRUE}

\item{na.rm}{A logical indicating whether NA:s should be removed from the
data. The default value is TRUE}
}
\value{
A matrix with as many columns equal to the number of columns in x
plus one for the weights, and rows with the inappropriate oobservations
removed.
}
\description{
This utility us used to tidy up or provide warnings in the functions that
use weights
}
\examples{

x <- c(NA, 1:5, Inf, -1, NaN, -Inf)
w <- rpois(length(x), 2)
clean_income(x, w)
x <- cbind(x, y = rpois(length(x), 10))
clean_income(x, w)

}
\seealso{
\code{\link{complete.cases}}, \code{\link{na.omit}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
