% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp_groups}
\alias{decomp_groups}
\title{Decomposition of inequality index by population subgroups}
\usage{
decomp_groups(
  x,
  w = rep(1, length(x)),
  eta = 2,
  data = NULL,
  ranked = x,
  na.rm = TRUE,
  group = NULL,
  func = "ge",
  ...
)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated, consisting of
income less some reference value}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the inequality aversion parameter. Larger value indicate higher
aversion, zero inequality neutrality and negative values inequality
preference.}

\item{data}{the data frame where the variables are found.}

\item{ranked}{the variable by which observations should be ranked (has no meaning for GE based within/between rankings)}

\item{na.rm}{a logical indicating whether NA's should removed.}

\item{group}{the (one-way) grouping varaiable}

\item{func}{the inequality index}
}
\value{
a vector with overall inequality, the within and the between term
}
\description{
This function decomposes generalized entroy (GE) inequality
indices by mutually exclusive population subgroup into a
within- and between-group term.
}
\details{
None.
}
\examples{

td <- data.frame(y=exp(rnorm(500)), g=sample(letters[1:4], 500, replace=TRUE))
decomp_groups(y, group=g, data=td)
}
\seealso{
\code{\link{ge}}
 @references Lambert, P. (1993).  \emph{The distribution and redistribution
of income. A mathematical analysis.} Manchester University Press,
Manchester.
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
