% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt.R
\name{fgt}
\alias{fgt}
\title{The Foster Greer Thorbecke poverty index}
\usage{
fgt(
  x,
  w = rep(1, length(x)),
  alpha = 0,
  mean = FALSE,
  fraction = 0.5,
  z = ifelse(mean, fraction * weighted_mean(x, w), fraction * weighted_median(x, w)),
  data = NULL,
  ranked = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{x}{the resource variable}

\item{w}{the sampling weights}

\item{alpha}{the FGT parameter}

\item{mean}{a logical indicating whether the default poverty line relate to
the mean rather than the median (which is the default)}

\item{fraction}{the fraction of the mean or median to use as the poverty
line (default is 0.5)}

\item{z}{the poverty line (defaults to the product of median and fraction))}

\item{data}{a data frame holding the data that contain x and w}

\item{na.rm}{a logical indicating whether NA's should removed}
}
\value{
A list with elements \item{fgt}{the index value} \item{n}{sample
size} \item{z}{poverty line} \item{sumq}{sum of weights}
}
\description{
This function estimates the Foster Greer Thorbecke poverty index for a
resource variable x and optional weights w.
}
\details{
This function estimates the Foster-Greer-Thorbecke poverty index for a
poverty line z. The poverty line z is taken by default to be 1/2 of the
weighted median, but can be any fraction of the median or mean or a value
given as an argument. For \eqn{\alpha=0}, FGT is the proportion of units
with income below z \eqn{\Pr(x < z)}. For \eqn{\alpha=1}, FGT equals
\eqn{\Pr(x < z)(1 - \bar x/z )} and for \eqn{\alpha = 2} FGT equals
\eqn{\Pr(x < z)(1 - \bar x/z )CV^2}, where \eqn{CV^2} is the squared
coefficient of variation.

The poverty line can be set explicitly or can be taken to be a fraction of
either the median (default) or the mean.
}
\examples{

fgt(runif(100))
n <- 100
income <- exp(10 + rnorm(n)*sqrt(2))
weight <- rpois(n,7)
fgt(income, weight)
sapply(seq(.3, .6, by = .1), function(x) fgt(income, weight, fraction = x)$fgt)

}
\references{
\insertRef{fostergreerandthorbecke1984}{incdist}
}
\seealso{
\code{\link{poverty}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
