% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{fields}
\alias{fields}
\title{Fields mobility index}
\usage{
fields(x, w = rep(1, dim(x)[1]), data = list(), na.rm = TRUE)
}
\arguments{
\item{x}{A matrix with income data}

\item{w}{A vector of weights}

\item{data}{A data frame in which to look for the data.}

\item{na.rm}{A logical, indicating whether or not to remove NAs}
}
\value{
A scalar.
}
\description{
This function estimates the Fields income mobility index.
}
\examples{


}
\references{
\insertRef{fieldsandok1999}{incdist}
}
\seealso{

}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
