% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge.R
\name{ge}
\alias{ge}
\alias{ge0}
\alias{ge1}
\alias{theil}
\alias{cv2}
\title{Generalized entropy inequality index}
\usage{
ge(x, w = rep(1, length(x)), eta = 2, data = NULL, ranked = x,
  na.rm = TRUE, ...)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the inequality aversion parameter, \eqn{\eta \ge 0}.  Larger
values place more weight on the inequality of high incomes.}

\item{data}{the data frame where the variables are found.}

\item{na.rm}{a logical indicating whether NA's should removed.}
}
\value{
The inequality index.
}
\description{
This function estimates members of the Generalized Entropy family of
inequality indices.
}
\details{
If \eqn{\eta = 0}, ge() estimates Theils 0 index, for \eqn{\eta = 1} ge()
estimate Thiels log mean difference. If \eqn{\eta = 2}, ge is 1/2 of the
square coefficient of variation, \eqn{\frac{\sigma^2_x}{2\mu_x}}.
}
\examples{

ge(runif(100), eta = 2)

}
\references{
Lambert, P. (1993).  \emph{The distribution and redistribution
of income. A mathematical analysis.} Manchester University Press,
Manchester.
}
\seealso{
\code{\link{gini}} \code{\link{atkinson}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
