% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\alias{gini.default}
\alias{gini.incdist}
\alias{concentration}
\alias{coefficient}
\title{Generalized gini or concentration coefficient}
\usage{
gini(x, ...)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated.}

\item{object}{an incdist object.}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the generalization parameter. For the default value of
eta = 2, gini() yields the ordinary gini coefficient.}

\item{na.rm}{a logical indicating whether NA's should removed.}

\item{data}{the data frame where the variables are found.}

\item{ranked}{an optional variable whch defines the ranking. If not
the default, the resulting index is a concentration
coefficient.}
}
\value{
The generalized gini or concentration coefficient.
}
\description{
This function estimates the generalized Gini coefficient or, if an
other variable is used for ordering, the concentration coefficient
of a variable with respect to the ordering variable. For the
generalizing parameter eta = 2, this yields the ordinary Gini
coefficient.
}
\details{
The function gini.default works on vectors, whereas the
gini.incdist function works on incdist objects (and currently
produces lots of output, much of which is unlikely to be useful.)
}
\examples{

gini(runif(20), eta = 2)
gini(runif(20), rpois(20, 5), eta = 2)

}
\references{
Lambert, P. (1993).  \emph{The distribution and
    redistribution of income. A mathematical analysis.} Manchester
    University Press, Manchester.
}
\seealso{
\code{\link{ineq}}, \code{\link{incdist}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
