% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{hasse}
\alias{hasse}
\title{Plot the Hasse diagram of a list of Lorenz or TIP curves}
\usage{
hasse(object, above.p = 0.1, lor.type = "ord", ...)
}
\arguments{
\item{object}{A list of Lorenz curve objects.}

\item{above.p}{Start comparison of this population fraction.}

\item{lor.type}{For lorenz curves, gives the type of Lorenz curve.}

\item{...}{Additional argumens to dominates.lorenz or dominates.tip.}
}
\value{
A plot.
}
\description{
hasse plots the Hasse diagram of a list of (ordinary, generalized or
absolute) Lorenz or TIP curves
}
\details{
A Hasse diagram is a vizualization of the strict partial order generated by
the Lorenz curve comparisons.
}
\examples{

require(relations)
lor.1 <- lorenz(runif(100))
lor.2 <- lorenz(rexp(100))
hasse(list("1"=lor.1, "2"=lor.2))


}
\seealso{
\code{\link{lorenz}}, \code{\link{tip}}.
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
