% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incdist.R
\name{inequality}
\alias{inequality}
\alias{inequality.incdist}
\alias{inequality.default}
\title{Estimate inequality indices}
\usage{
inequality(x, ...)
}
\arguments{
\item{x}{a numerical vector or incdist object}

\item{w}{an optional vector of non-negative integer values weights.}

\item{indices}{to estimate}
}
\value{
the estimated indices
}
\description{
This function estimates inequality indices.
}
\references{
Lambert, P. (1993).  \emph{The distribution and
    redistribution of income. A mathematical analysis.} Manchester
    University Press, Manchester.
}
\seealso{
\code{\link{gini}}, \code{\link{ge}}, \code{\link{atkinson}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
