% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ineqvar}
\alias{ineqvar}
\title{Variance of inequality index
based on influence}
\usage{
ineqvar(x, w, ranked = x, na.rm = FALSE, index = "gini", ...)
}
\arguments{
\item{x}{an income vector or matrix}

\item{w}{the vector of weights}

\item{ranked}{variable by which data are ranked}

\item{na.rm}{A logical indicating whether NA:s should be removed from the
data. The default value is TRUE}

\item{index}{the index in question, one of gini, ge2, ge0, ge1}
}
\description{
Variance of inequality index
based on influence
}
\references{
\insertRef{cowellandflachaire2015}{incdist}
}
