% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolm.R
\name{kolm}
\alias{kolm}
\title{Kolm's inequality index}
\usage{
kolm(x, w = rep(1, length(x)), eta = 1, data = NULL,
  absolute = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the inequality aversion parameter (>0). Larger value indicate
higher aversion, zero inequality neutrality.}

\item{data}{the data frame where the variables are found.}

\item{absolute}{a logical, indicating if the absolute rather than relative
measure be given.}

\item{na.rm}{a logical indicating whether NA's should removed.}
}
\description{
This function estimates the inequality indices (relative or absolute) of
Kolm.
}
\details{
The inequality index.
}
\examples{

kolm(runif(100), eta = 1)

}
\references{
Lambert, P. (1993).  \emph{The distribution and redistribution
of income. A mathematical analysis.} Manchester University Press,
Manchester.
}
\seealso{
\code{\link{gini}} \code{\link{ge}}, \code{\link{atkinson}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
