% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.R
\name{plot.lorenz}
\alias{plot.lorenz}
\alias{lines.lorenz}
\alias{plot.lorenz_list}
\title{Plot a Lorenz curve}
\usage{
\method{plot}{lorenz}(
  x,
  add = FALSE,
  lor.type = "ord",
  diff = FALSE,
  xlab = FALSE,
  ylab = FALSE,
  ylim = FALSE,
  xlim = FALSE,
  type = "l",
  lty = FALSE,
  ...
)
}
\arguments{
\item{x}{A Lorenz curve object.}

\item{add}{A logical. If TRUE, the Lorenz curve is added to a plot.}

\item{lor.type}{The type of Lorenz curve to be drawn. If "ord", the graph is
an ordinary Lorenz curve \eqn{L(p)}, if "gen", the graph is a generalized
Lorenz curve \eqn{\mu_x L(p)}.}

\item{diff}{A logical. If TRUE, graph is of \eqn{p - L(p)} rather than
\eqn{L(p)}.}

\item{xlab}{The x axis label. Defaults to "Population share".}

\item{ylab}{The y axis label. Defaults to "Income share".}

\item{...}{Additional graphics arguments to pass to the plot function.}
}
\value{
A plot.
}
\description{
plot.lorenz plots a (generalized) Lorenz curve.
}
\details{
plot.lorenz_list plots a list of lorenz curve on the same graph and optionally
places a legend using the names of the lorenz curves as labels.

An ordinary Lorenz curve is bounded in the unit box, is positive and
increasing and lies everywhere below the 45 degree line. By default the
function draws the 45 degree line and sets ylim and xlim to c(0,1).

If several generalized Lorenz curves are drawn, or ordinary Lorenx curves
are drawn with diff = FALSE, care need to be taken that the ylim gets
correectly set.
}
\examples{

plot(lorenz(runif(100), q = 5))

}
\references{

}
\seealso{
\code{\link{lorenz}}.
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
