% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.R
\name{plot.tip}
\alias{plot.tip}
\alias{lines.tip}
\alias{plot.tipl}
\title{Plot a TIP or CPG curve}
\usage{
\method{plot}{tip}(object, add = FALSE, xlab = FALSE, ylab = FALSE,
  xlim = FALSE, ylim = FALSE, type = FALSE, ...)
}
\arguments{
\item{object}{An object of class tip.}

\item{add}{A logical. If TRUE, the curve is added to an existing plot.}
}
\value{
A plot.
}
\description{
plot.tip plots a so-called "Three I's of Poverty" (TIP) or Cumulative
Poverty Gap (CPG) object.
}
\details{
lines.tip add lines to an existing tip plot.

plot.tipl plots a list of tip curves.

Plots the TIP curve.
}
\examples{

plot(tip(runif(100)))

}
\references{
Jenkins, Stephen P & Peter J Lambert ( 1997), `Three `i's of
poverty curves, with an analysis of U.K. poverty trends', Oxford Economic
Papers 49(3), 317-327.
}
\seealso{
\code{\link{tip}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
