% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incdist.R
\name{poverty}
\alias{poverty}
\alias{povery}
\alias{poverty.incdist}
\alias{poverty.default}
\title{Estimate poverty indices}
\usage{
poverty(x, ...)
}
\arguments{
\item{x}{a numerical vector or incdist object}

\item{w}{an optional vector of non-negative integer values weights.}

\item{indices}{to estimate}
}
\value{
the estimated indices
}
\description{
This function estimates poverty indices.
}
\references{
Lambert, P. (1993).  \emph{The distribution and
    redistribution of income. A mathematical analysis.} Manchester
    University Press, Manchester.
}
\seealso{
\code{\link{inequality}}, \code{\link{fgt}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
