% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect.R
\name{prospect}
\alias{prospect}
\alias{ede}
\title{Inequality index based on prospect theory}
\usage{
prospect(x, w = rep(1, length(x)), eta = 1, data = NULL,
  na.rm = TRUE)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated, consisting of
income less some reference value}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the inequality aversion parameter. Larger value indicate higher
aversion, zero inequality neutrality and negative values inequality
preference.}

\item{data}{the data frame where the variables are found.}

\item{na.rm}{a logical indicating whether NA's should removed.}
}
\value{
The inequality index.
}
\description{
This function estimates Atkinson's inequality index.
}
\details{
None.
}
\examples{

prospect(runif(100), eta = 1/2)

}
\references{
Lambert, P. (1993).  \emph{The distribution and redistribution
of income. A mathematical analysis.} Manchester University Press,
Manchester.
}
\seealso{
\code{\link{gini}} \code{\link{ge}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
