% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{shorrocks}
\alias{shorrocks}
\title{Shorrocks mobility index}
\usage{
shorrocks(x, w = rep(1, dim(x)[1]), data = NULL, na.rm = TRUE,
  index = "gini", ...)
}
\arguments{
\item{x}{A matrix with income data}

\item{w}{A vector of weights}

\item{data}{A data frame in which to look for the data}

\item{na.rm}{A logical, indicating whether or not to remove NAs}

\item{index}{An inequality index.}
}
\value{
A scalar.
}
\description{
This function estimates the Shorrocks income mobility index for a specific
inequality indext
}
\examples{


}
\references{
Cite Shorrocks's Econometrica paper
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
