% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{tip}
\alias{tip}
\alias{tip.default}
\alias{cpg}
\alias{tip.locfit}
\alias{tip.incdist}
\alias{is.}
\alias{summary.tip}
\alias{print.summary.tip}
\alias{dominates.tip}
\alias{dominates.tip_list}
\title{The "three I's of poverty" (TIP) or "Cumulated Poverty Gap" (CGP) curve}
\usage{
tip(x, ...)
}
\arguments{
\item{x}{the resource variable}

\item{w}{the sampling weights}

\item{q}{an integer or logical. If an integer, it is the number of equally
spaced points at which the Lorenz curve ordinates are estimated. If FALSE,
then the Lorenz curve is estimated at every point in x.}

\item{z}{the poverty line}

\item{data}{the data frame where the variables are found.}

\item{alpha}{the FGT parameter}

\item{normalize}{a logical. If true, tip returns the normalized gap (i.e.,
the gap divided by the poverty line).}

\item{mean}{a logical indicating whether the default poverty line relate to
the mean rather than the median}

\item{fraction}{the fraction of the mean or median to use as the poverty
line}

\item{na.rm}{a logical indicating whether NA's should removed}
}
\value{
A object of class tip with elements \item{p}{the population
proportion} \item{ordinates}{the TIP curve value at p} \item{hc}{the
proportion of units with x less than z} \item{relgap}{the average poverty
gap (across the poor population)} \item{n}{sample size} \item{z}{the poverty
line} \item{sum.weights}{the sum of weights}
}
\description{
This function returns the cumulated poverty gap.
}
\details{
This function computes the `TIP`curve, that is, the cumulated poverty gap as
a function of the population proportion as \deqn{TIP(p) := \sum_{F^{-1}(x_i)
< p} I(x_i < z) (z - x_i),} or, if the normalized by the poverty line,
\deqn{TIP(p) := \sum_{F^{-1}(x_i) < p} I(x_i < z) (1 - x_i/z).} The TIP
curve is increasing for values of x below the poverty line z, at which point
it becomes horizontal. That point gives the proportion below the poverty
line (p) and the cumulative average poverty gap (the value of the TIP
curve).  If weights are given, the function returns the weighted TIP curve.
If \eqn{q = FALSE}, tip() returns an object whose elements \eqn{p} and
\eqn{ordinates} are of length(x).

Simple summary, print and plot methods are provided.
}
\examples{

income <- rexp(100)
weight <- rpois(100,5)
tip.ex <- tip(income, weight, q = FALSE)
plot(tip.ex, lwd = 2)

}
\references{
\insertRef{jenkinsandlamber1997}{incdist}
}
\seealso{
\code{\link{lorenz}}, \code{\link{fgt}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
