% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{weighted_moment}
\alias{weighted_moment}
\alias{weighted_var}
\alias{weighted_std}
\alias{weighted_quantile}
\alias{weighted_median}
\alias{weighted_sum}
\title{Weighted statistic}
\usage{
weighted_moment(x, w, a = 1, ranked = x, na.rm = FALSE)
}
\arguments{
\item{x}{the variable whose moment is to be estimated}

\item{w}{the weights}

\item{a}{the degree of the moment}

\item{na.rm}{a logical indicating whether NA's should removed}

\item{probs}{the probabilities of the quantiles}

\item{names}{no idea what this does}
}
\value{
The desired statistic. Weighted quantile produces a vector with
length(probs).
}
\description{
The weighted_moment is a wrapper around R's weighted_mean which raises the
variable x to the power a before taking the weighted mean.  weighted_var and
weighted_std are based on weighted_moment.
}
\details{
The weighted_quantile function yields a version of quantiles which does not
quite reach the sophistication of that built into R. In particular, for a
probability p this function yields the highest x for which cumsum(w)/sum(w)
< p, rather than some interpolated value.  weighted_median is
weighted_quantile(x,w, p = .5)
}
\examples{

x <- rexp(100)
w <- rpois(100,5)
weighted_median(x, w)
weighted_moment(x,w, 1)
weighted_mean(x,w, 1)
weighted_var(x,w)

}
\seealso{
\code{\link{weighted_mean}}, \code{\link{cov.wt}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
