% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\alias{gini.default}
\alias{gini.incdist}
\alias{concentration_coefficient}
\title{Generalized gini or concentration coefficient}
\usage{
gini(x, ...)
}
\arguments{
\item{x}{a numerical vector whose index is to be estimated.}

\item{object}{an incdist object.}

\item{w}{an optional vector of non-negative integer values weights.}

\item{eta}{the generalization parameter. For the default value of
eta = 2, gini() yields the ordinary gini coefficient.(Makes little sense to type = "absolute".)}

\item{na.rm}{a logical indicating whether NA's should removed.}

\item{data}{the data frame where the variables are found.}

\item{ranked}{an optional variable whch defines the ranking. If not
the default, the resulting index is a concentration
coefficient.}

\item{no.negatives}{indicates whether or not negative incomes should be
allowed.}

\item{type}{whether to estimate an absolute or relative gini.}

\item{absolute}{alternative for type = "absolute".}

\item{small}{if set to TRUE estimates the gini coefficient by integrating a  micro-data-based Lorenz curve. (This ensures the resulting gini respects the Lorenz order)}
}
\value{
The generalized gini or concentration coefficient.
}
\description{
This function estimates the generalized Gini coefficient or, if an
other variable is used for ordering, the concentration coefficient
of a variable with respect to the ordering variable. For the
generalizing parameter eta = 2, this yields the ordinary Gini
coefficient.
}
\details{
The function gini.default works on vectors, whereas the
gini.incdist function works on incdist objects (and currently
produces lots of output, much of which is unlikely to be useful.)
}
\examples{

gini(runif(20), eta = 2)
gini(runif(20), rpois(20, 5), eta = 2)

}
\references{
\insertRef{lambert1993}{incdist} \insertRef{davidson2009}{incdist}
}
\seealso{
\code{\link{inequality}}, \code{\link{incdist}}, \code{\link{lorenz}}
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
