% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{hasse}
\alias{hasse}
\title{Plot the Hasse diagram of a list of Lorenz or TIP curves}
\usage{
hasse(object, above.p = 0.1, lor.type = "ord", ...)
}
\arguments{
\item{object}{A list of Lorenz curve objects.}

\item{above.p}{Start comparison of this population fraction.}

\item{lor.type}{For lorenz curves, gives the type of Lorenz curve.}

\item{...}{Additional arguments to dominates.lorenz or dominates.tip.}
}
\value{
A plot.
}
\description{
hasse plots the Hasse diagram of a list of (ordinary, generalized or
absolute) Lorenz or TIP curves
}
\details{
A Hasse diagram is a vizualization of the strict partial order generated by
the Lorenz curve comparisons.
}
\examples{

lor.1 <- lorenz(rep(c(3, 3, 3, 3, 3), each=20), q = FALSE)
lor.2 <- lorenz(rep(c(1, 2, 3, 4, 5), each=20), q = FALSE)
lor.3 <- lorenz(rep(c(1, 1, 3, 4, 6), each=20), q = FALSE)
lor.4 <- lorenz(rep(c(1.5, 2, 3, 4, 5.5), each=20), q = FALSE)
lor.5 <- lorenz(rep(c(1, 1, 1, 5, 7), each=20), q = FALSE)
hasse(list("a"=lor.1, "b"=lor.2, "c"=lor.3, "d"=lor.4, "e"=lor.5))


}
\references{
\insertRef{atkinsonsmeedingandrainwater1995}{incdist}
}
\seealso{
\code{\link{lorenz}}, \code{\link{tip}}.
}
\author{
Markus Jantti \email{markus.jantti@iki.fi}
}
